refactor into sub-modules; implement constructors
- implements a system of partial constructors whereby you can compose
  transformations into structs by writing functions that return
  named tuples with the members of the struct that you added in this
  component
- refactors different syntax passes into modules, resulting in cleaner
  code
- refactor tests into independent test sets
