+++ test/runtests.jl
@@ -1,211 +1,11 @@
 # ~/~ begin <<docs/src/50-implementation.md#test/runtests.jl>>[init]
 using Test
-using ModuleMixins:
-    @spec,
-    @spec_mixin,
-    @spec_using,
-    @mixin,
-    Struct,
-    parse_struct,
-    define_struct,
-    Pass,
-    @compose,
-    @for_each
-using MacroTools: prewalk, rmlines
 
-clean(expr) = prewalk(rmlines, expr)
-
-# ~/~ begin <<docs/src/50-implementation.md#test-toplevel>>[init]
-@spec module MySpec
-const msg = "hello"
-end
-# ~/~ end
-# ~/~ begin <<docs/src/50-implementation.md#test-toplevel>>[1]
-@spec_mixin module MyMixinSpecOne
-@mixin A
-end
-@spec_mixin module MyMixinSpecMany
-@mixin A, B, C
-end
-# ~/~ end
-# ~/~ begin <<docs/src/50-implementation.md#test-toplevel>>[2]
-@spec_using module SU_A
-const X = :hello
-export X
-end
-
-@spec_using module SU_B
-@mixin SU_A
-const Y = X
-end
-
-@spec_using module SU_C
-const Z = :goodbye
-end
-
-@spec_using module SU_D
-@mixin SU_A
-@mixin SU_B, SU_C
-end
-# ~/~ end
-# ~/~ begin <<docs/src/50-implementation.md#test-toplevel>>[3]
-struct EmptyPass <: Pass
-    tag::Symbol
-end
-# ~/~ end
-# ~/~ begin <<docs/src/50-implementation.md#test-toplevel>>[4]
-module ComposeTest1
-using ModuleMixins
-
-@compose module A
-    struct S
-        a::Int
-    end
-end
-
-@compose module B
-    struct S{T}
-        b::T
-    end
-end
-
-@compose module AB
-    @mixin A, B
-end
-end
-# ~/~ end
-# ~/~ begin <<docs/src/50-implementation.md#test-toplevel>>[5]
-module Common
-    export AbstractData
-    abstract type AbstractData end
-end
-
-@compose module WriterA
-    using ..Common
-
-    @kwdef struct Data <: AbstractData
-        a::Int
-    end
-
-    function write(io::IO, data::AbstractData)
-        println(io, data.a)
-    end
-end
-
-@compose module WriterB
-    using ..Common
-    @mixin WriterA
-
-    @kwdef struct Data <: AbstractData
-        b::Int
-    end
-
-    function write(io::IO, data::AbstractData)
-        println(io, data.b)
-    end
-end
-
-@compose module WriterC
-end
-
-@compose module WriterABC
-    using ModuleMixins
-    @mixin WriterB, WriterC
-
-    function write(io::IO, data::AbstractData)
-        @for_each(P->P.write(io, data), PARENTS)
-    end
-end
-# ~/~ end
-
-@testset "ModuleMixins" begin
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[init]
-    @testset "@spec" begin
-        @test clean.(MySpec.AST) == clean.([:(const msg = "hello")])
-        @test MySpec.msg == "hello"
-    end
-    # ~/~ end
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[1]
-    @testset "@spec_mixin" begin
-        @test MyMixinSpecOne.PARENTS == [:A]
-        @test MyMixinSpecMany.PARENTS == [:A, :B, :C]
-    end
-    # ~/~ end
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[2]
-    @testset "@spec_using" begin
-        @test SU_B.Y == SU_A.X
-        @test SU_B.PARENTS == [:SU_A]
-        @test SU_D.PARENTS == [:SU_A, :SU_B, :SU_C]
-        @test SU_D.SU_C.Z == :goodbye
-    end
-    # ~/~ end
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[3]
-    @testset "pass composition" begin
-        a = EmptyPass(:a) + EmptyPass(:b)
-        @test a.parts[1].tag == :a
-        @test a.parts[2].tag == :b
-    end
-    # ~/~ end
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[4]
-    cases = Dict(
-        :(struct A x end) => Struct(false, false, :A, nothing, nothing, [:x]),
-        :(mutable struct A x end) => Struct(false, true, :A, nothing, nothing, [:x]),
-        :(@kwdef struct A x end) => Struct(true, false, :A, nothing, nothing, [:x]),
-        :(@kwdef mutable struct A x end) => Struct(true, true, :A, nothing, nothing, [:x]),
-        :(struct A{T} x::T end) => Struct(false, false, :A, [:T], nothing, [:(x::T)]),
-    )
-    
-    for (k, v) in pairs(cases)
-        @testset "Struct mangling: $(join(split(string(clean(k))), " "))" begin
-            @test clean(define_struct(parse_struct(k))) == clean(k)
-            @test clean(define_struct(v)) == clean(k)
-        end
-    end
-    # ~/~ end
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[5]
-    @testset "Struct mangling abstracts" begin
-        @test parse_struct(:(struct A <: B x end)).abstract_type == :B
-        @test parse_struct(:(mutable struct A <: B x end)).abstract_type == :B
-    end
-    
-    @testset "Mangling type arguments" begin
-        using ModuleMixins: mangle_type_parameters!
-        let s = parse_struct(:(struct S{T} x::T end))
-            @test s.type_parameters == [:T]
-            mangle_type_parameters!(s, :A)
-            @test s.type_parameters == [:_T_A]
-            @test clean(define_struct(s)) == clean(:(struct S{_T_A} x::_T_A end))
-        end
-        let s = parse_struct(:(struct S{T} x::Vector{T} = [] end))
-            mangle_type_parameters!(s, :A)
-            @test clean(define_struct(s)) == clean(:(struct S{_T_A} x::Vector{_T_A} = [] end))
-        end
-    end
-    # ~/~ end
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[6]
-    @testset "compose struct members" begin
-        @test ComposeTest1.AB.PARENTS == [:A, :B]
-        @test fieldnames(ComposeTest1.AB.S) == (:a, :b)
-    end
-    @testset "compose hierarchy" begin
-        @test ComposeTest1.AB.MIXIN_TREE == IdDict(:AB => [:A, :B], :A => [], :B => [])
-        @test WriterABC.MIXIN_TREE == IdDict(
-            :WriterABC => [:WriterB, :WriterC],
-            :WriterC => [],
-            :WriterB => [:WriterA],
-            :WriterA => [])
-    end
-    @testset "composed struct has type parameter" begin
-        @test ComposeTest1.AB.S{Float64}(1, 2).b isa Float64
-    end
-    # ~/~ end
-    # ~/~ begin <<docs/src/50-implementation.md#test>>[7]
-    @testset "for-each" begin
-        io = IOBuffer(write=true)
-        data = WriterABC.Data(a = 42, b = 23)
-        WriterABC.write(io, data)
-        @test String(take!(io)) == "23\n42\n"
-    end
-    # ~/~ end
+@testset "ModuleMixins.jl" begin
+    include("SpecSpec.jl")
+    include("PassesSpec.jl")
+    include("MixinsSpec.jl")
+    include("StructsSpec.jl")
+    include("ComposeSpec.jl")
 end
 # ~/~ end
